Tests/LibWeb/Text/expected/UIEvents/pointerdown-prevent-default.txt

> click with pointerdown preventDefault
pointerdown
pointerup
click
